pyproject.toml
[tool.poetry]
name = "back-templaria"
version = "0.1.0"
description = ""
authors = ["Gabriel Vilarino <gabriel_vilarino@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.2.1"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.161.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
